cast it directly to integer

+++ src/de/Keyle/MyWolf/skill/MyWolfExperience.java
@@ -151,7 +151,7 @@ public class MyWolfExperience
             ScriptEngine se = parseJS();
             try
             {
-                return ((Double) se.get("lvl")).intValue();
+                return (Integer)se.get("lvl");
             }
             catch (Exception e)
             {
